Remove bindings to resize event
We had an issue with this event: the dashboard view stays in memory,
and on every resize (which includes software keyboard opening /
closing) carousel recalculated the size of the slides.
Since we don't need to react to the resize event, solving the issue
seems to be as easy as commenting out the event binding.